+++ realm-transformer/src/main/kotlin/io/realm/transformer/ext/ProjectExt.kt
@@ -21,17 +21,56 @@ import org.gradle.api.Project
 import java.io.File
 
 /**
+ * Attempts to determine the best possible unique AppId for this project.
+ */
+fun Project.getAppId(): String {
+    // Use the Root project name, usually set in `settings.gradle`
+    // This means that we don't treat apps with multiple flavours as different, nor
+    // if a project contains more than one app (probably unlikely).
+    // This seems acceptable. These cases would just show up as more builds for the
+    // same AppId.
+    return this.rootProject.name
+}
+
+/**
  * Returns the `targetSdk` property for this project if it is available.
  */
-fun Project.getTargetSdk(): String? {
-    return getAndroidExtension(this).defaultConfig?.targetSdkVersion?.apiString
+fun Project.getTargetSdk(): String {
+    return getAndroidExtension(this).defaultConfig.targetSdkVersion?.apiString ?: "unknown"
 }
 
 /**
  * Returns the `minSdk` property for this project if it is available.
  */
-fun Project.getMinSdk(): String? {
-    return getAndroidExtension(this).defaultConfig?.minSdkVersion?.apiString
+fun Project.getMinSdk(): String {
+    return getAndroidExtension(this).defaultConfig.minSdkVersion?.apiString ?: "unknown"
+}
+
+/**
+ * Returns the version of the Android Gradle Plugin that is used.
+ */
+fun Project.getAgpVersion(): String {
+    // This API is only available from AGP 7.0.0. And it is a bit unclear exactly which part of
+    // this is actually stable. Also, there appear to be problems with depending on AGP 7.* on
+    // the compile classpath (it cannot load BaseExtension).
+    //
+    // So for now, this code assumes that we are compiling against AGP 4.1 and uses reflection
+    // to try to grap the AGP version.
+    //
+    // This is done with a best-effort, but we just
+    // accept finding the version isn't possible if anything goes wrong.
+    return try {
+        val extension = this.extensions.getByName("androidComponents") as Object
+        val method = extension.`class`.getMethod("getPluginVersion")
+        val version = method.invoke(extension)
+        if (version != null) {
+            return version.toString()
+        } else {
+            return "unknown"
+        }
+    } catch (e: Exception) {
+        "unknown"
+    }
 }
 
 /**